+++ test/net/lewisship/cli_tools/aux.clj
@@ -29,3 +29,23 @@
                                             (reset! *errors# errors#))]
        (with-exit 1 ~@body))
      (is (= @*errors# ~expected-errors))))
+
+(defmacro capture-result
+  "Executes body and capture standard out, standard error, and exit status.
+  :out :err and :status."
+  [& body]
+  `(let [out#    (StringWriter.)
+         err#    (StringWriter.)
+         status# (binding [*out* out#
+                           *err* err#]
+                   (try
+                     (do ~@body)
+                     ;; Normal exit is a 0
+                     0
+                     (catch Throwable t#
+                       (when-not (= "Exit" (ex-message t#))
+                         (throw t#))
+                       (-> t# ex-data :status))))]
+     {:out    (str out#)
+      :err    (str err#)
+      :status status#}))